[IMP]sale_order_line_position: move order line position to a mixin
