+++ sale_order_line_position/models/__init__.py
@@ -1,4 +1,5 @@
 from . import ir_actions_report
+from . import position_mixin
 from . import sale_order
 from . import sale_order_line
 from . import res_company
